requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0  # 用于加载 .env 配置文件

# LLM支持(可选 - 如需使用ECNU API功能,请取消注释并安装)
# openai==1.3.0  # 使用ECNU API时需要（ECNU API使用OpenAI兼容接口）

# 数据导出支持(必需)
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
